+++ @commitlint/config-conventional/index.js
@@ -36,7 +36,7 @@ module.exports = {
 	prompt: {
 		questions: {
 			type: {
-				description: "Select the type of change that you're committing:",
+				description: "Select the type of change that you're committing",
 				enum: {
 					feat: {
 						description: 'A new feature',